add some tests

+++ mix.exs
@@ -26,17 +26,21 @@ defmodule Ci.MixProject do
   end
 
   defp gen_boot(_) do
-    Mix.shell().info("Generating node.rel file")
+    if not File.exists?("priv/node.rel") do
+      Mix.shell().info("Generating node.rel file")
 
-    apps =
-      [:kernel, :stdlib, :elixir, :compiler]
-      |> Enum.map(&{&1, get_version(&1)})
+      apps =
+        [:kernel, :stdlib, :elixir, :compiler]
+        |> Enum.map(&{&1, get_version(&1)})
 
-    rel_spec = {:release, {'node', '0.1.0'}, {:erts, :erlang.system_info(:version)}, apps}
-    File.write!("priv/node.rel", consultable(rel_spec))
+      rel_spec = {:release, {'node', '0.1.0'}, {:erts, :erlang.system_info(:version)}, apps}
+      File.write!("priv/node.rel", consultable(rel_spec))
+    end
 
-    Mix.shell().info("Generating node.boot file")
-    :systools.make_script('priv/node', [:silent])
+    if not File.exists?("priv/node.rel") do
+      Mix.shell().info("Generating node.boot file")
+      :systools.make_script('priv/node', [:silent])
+    end
   end
 
   defp consultable(term) do
